// build-test/src/types/request.d.ts
export declare namespace RequestParams {
// ...
  interface CreateAndUpdateAppList {
    code_language: string | null;
    git_url: string;
    id: number | null;
    name: string;
    remark: string;
    run_env: number | null;
  }
// ...
}